requirements.txt
chromadb>=0.4.0
edge-tts>=6.1.9
faster-whisper>=0.10.0
google-generativeai>=0.8.5
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.1.0
sentence-transformers>=2.2.2
streamlit>=1.40.0
